+++ src/libAtomVM/globalcontext.h
@@ -182,6 +182,7 @@ struct GlobalContext
 #ifdef HAVE_PSA_CRYPTO
     ErlNifResourceType *psa_hash_op_resource_type;
     ErlNifResourceType *psa_cipher_op_resource_type;
+    ErlNifResourceType *psa_mac_op_resource_type;
 #endif
 
     void *platform_data;